Added - Swap headgear for slung helmet interaction
What
=================
Allow the player to swap an already slung helmet with the headgear they
are currently wearing. If the headgear is not whitelisted to be slung,
then add the headgear to the player's inventory or the ground.

Why
=================
So that the player doesn't have to mess around with their inventory in
order to swap out the headgear.

Issue ID: N/A